ROX-30110: Enable export-default-react lint rule with ignores
