+++ ui/apps/platform/package.json
@@ -101,8 +101,8 @@
         "build:ocp-plugin": "tsc && NODE_ENV=production webpack --config webpack.ocp-plugin.config.js",
         "test": "TZ=UTC vitest",
         "test-coverage": "TZ=UTC vitest run --coverage",
-        "lint": "eslint --quiet .",
-        "lint:fix": "eslint --fix --quiet .",
+        "lint": "eslint --quiet",
+        "lint:fix": "eslint --fix --quiet",
         "tsc": "tsc",
         "cypress-open": "./scripts/cypress.sh open --e2e --config defaultBrowser=chrome",
         "cypress-spec": "./scripts/cypress.sh run --spec",
