+++ ui/apps/platform/src/Components/Popper.jsx
@@ -79,4 +79,7 @@ const CustomPopperContainer = (props) => {
     );
 };
 
+// Encapsulate CustomPotter and CustomPopperContainer as implementation details,
+// especially since Popper appears in Find results.
+// eslint-disable-next-line limited/export-default-react
 export default CustomPopperContainer;
